Split up files, add decode from data method
